refactor swimming fishes from tick to fish model
