+++ src/models/tick.js
@@ -1,4 +1,5 @@
-const { highestTick, deepCopy, randomPosOrNeg } = require('../utils');
+const { highestTick, deepCopy } = require('../utils');
+const { swimFishes } = require('./fish');
 
 const createTick = (dataStore, tick) => {
   const _highestTick = highestTick(dataStore.ticks);
@@ -29,33 +30,7 @@ const createTick = (dataStore, tick) => {
     }
 
     // All existing fishes should swim to a random point near them.
-    state.fishes = state.fishes.map((_fish) => {
-      const nextLocation = [
-        _fish.location[0] + randomPosOrNeg(4),
-        _fish.location[1] + randomPosOrNeg(4),
-        _fish.location[2] + randomPosOrNeg(4),
-      ];
-
-      // If the next location x coordinate would fall outside the bounding box of the aquarium,
-      // restrain the next fish location to its current one.
-      if (nextLocation[0] < dataStore.aquarium.minX || nextLocation[0] > dataStore.aquarium.maxX) {
-        nextLocation[0] = _fish.location[0];
-      }
-
-      // y.
-      if (nextLocation[1] < dataStore.aquarium.minY || nextLocation[1] > dataStore.aquarium.maxY) {
-        nextLocation[1] = _fish.location[1];
-      }
-
-      // z.
-      if (nextLocation[2] < dataStore.aquarium.minZ || nextLocation[2] > dataStore.aquarium.maxZ) {
-        nextLocation[2] = _fish.location[2];
-      }
-
-      _fish.location = nextLocation;
-
-      return _fish;
-    });
+    state.fishes = swimFishes(state.fishes, dataStore);
 
     // Is a fish spawned in the current tick?
     const fish = dataStore.fishes.filter(fish => fish.tick === id);
